+++ site/conf/evolutions/default/6.sql
@@ -2,8 +2,8 @@
 
 -- Relates case studies to collections m:n
 CREATE TABLE collection_study ( 
-	case_study_id INTEGER NOT NULL,
-	collection_id INTEGER NOT NULL,
+	case_study_id BIGINT NOT NULL,
+	collection_id BIGINT NOT NULL,
 	PRIMARY KEY(collection_id, case_study_id),
 	FOREIGN KEY(case_study_id) REFERENCES case_study(id),
 	FOREIGN KEY(collection_id) REFERENCES collection(id)
@@ -11,8 +11,8 @@ CREATE TABLE collection_study (
 
 -- Relates images to collections m:n
 CREATE TABLE collection_image (
-	collection_id INTEGER NOT NULL,
-	image_id INTEGER NOT NULL,
+	collection_id BIGINT NOT NULL,
+	image_id BIGINT NOT NULL,
 	notes VARCHAR(10240) NOT NULL,
 	damage_indicator INTEGER NOT NULL,
 	degree_of_damage INTEGER NOT NULL,
